Add preloading data on server and rehydration on the client
Renders twice on the server - once to fire off data loaders and once when
they've all finished so they can render with correct state. The pre loaded
data is then sent to the client as a global variable so it can be rehydrated.
Redux makes this easier by having a central state but this is a non-redux PoC.