Fix bug with paths being replace after content had been rendered
This meant we were transforming paths into /consultations for the rendered
html, including anchor tags. We only want to do it for anything that's inside
the static index.html file, and NOT what's rendered via react. We assume links
rendered via react on the server have the correct path because of the use of
basename within StaticRouter