EC-810 Add collection decorator to tests
Due to the usage of a shared context and database in tests
and some of the tests running database teardown and setup
methods there was an issue where some tests would interfere
with each other by running in parallel. We added the xunit
Collection decorator to each test class so they run sequentially.